画饼 网页传送

+++ Modules/MinecraftBEFlow/Resources/views/servers/show.blade.php
@@ -7,13 +7,18 @@
 @endsection
 
 @section('content')
+    @php($cache = cache('mcbe_flow_server_' . $server->id, 0))
+    @php($in_team = userInTeam($server->team_id))
+
     <div class="mdui-typo-display-1 server-name">{{ $server->name }}</div>
 
     <div class="mdui-tab mdui-m-b-2" mdui-tab>
         <a href="#infomations" class="mdui-ripple">{{ tr('Information') }}</a>
         <a href="#players" class="mdui-ripple">{{ tr('Online Players') }}</a>
-        <a href="#configuration" class="mdui-ripple">{{ tr('Basic Configuration') }}</a>
-        <a href="#delete" class="mdui-ripple">{{ tr('Delete') }}</a>
+        @if ($in_team)
+            <a href="#configuration" class="mdui-ripple">{{ tr('Basic Configuration') }}</a>
+            <a href="#delete" class="mdui-ripple">{{ tr('Delete') }}</a>
+        @endif
     </div>
 
     <div id="infomations">
@@ -21,13 +26,45 @@
             <p>Your server need active, please download "EdgeFlow.lxl.js" to your plugins folder, then restart server.
             </p>
         @else
-            @php($cache = cache('mcbe_flow_server_' . $server->id, 0))
             @if (!$cache)
                 Server offline
             @else
-                version: {{ $cache['version'] ?? 0 }}
+                <div class="mdui-table-fluid">
+                    <table class="mdui-table mdui-table-hoverable">
+                        <tbody class="mdui-typo">
+                            <tr>
+                                <td>{{ tr('Online Players') }}</td>
+                                <td>
+                                    {{ $cache['players_count'] ?? 0 }}
+                                </td>
+                            </tr>
+                            <tr>
+                                <td>{{ tr('Server Version') }}</td>
+                                <td>
+                                    {{ $cache['version'] ?? 'Unknown' }}
+                                </td>
+                            </tr>
+
+                            {{-- <tr>
+                                <td>{{ tr('Transfer to Server') }}</td>
+                                <td>
+                                    <a>Transfer</a>
+                                </td>
+                            </tr> --}}
+
+                            <tr>
+                                <td>{{ tr('Add Server') }}</td>
+                                <td>
+                                    <a
+                                        href="minecraft://?addExternalServer={{ $server->name }}|{{ ipPort($server->ip, $server->port) }}">{{ tr('Add Server') }}</a>
+                                </td>
+                            </tr>
+
+
+                        </tbody>
+                    </table>
+                </div>
             @endif
-            {{-- @if (cache()) --}}
         @endif
     </div>
     <div id="players">
@@ -50,46 +87,48 @@
 
     </div>
 
+    @if ($in_team)
 
-    <div id="configuration">
-        <form onsubmit="event.preventDefault();m.update(this)">
-            <div class="mdui-textfield mdui-textfield-floating-label">
-                <label class="mdui-textfield-label">{{ tr('Server Name') }}</label>
-                <input class="mdui-textfield-input" type="text" name="name" value="{{ $server->name }}"
-                    oninput="$('.server-name').text(this.value)" />
-                <div class="mdui-textfield-helper">{{ tr('Your server name.') }}</div>
-            </div>
+        <div id="configuration">
+            <form onsubmit="event.preventDefault();m.update(this)">
+                <div class="mdui-textfield mdui-textfield-floating-label">
+                    <label class="mdui-textfield-label">{{ tr('Server Name') }}</label>
+                    <input class="mdui-textfield-input" type="text" name="name" value="{{ $server->name }}"
+                        oninput="$('.server-name').text(this.value)" />
+                    <div class="mdui-textfield-helper">{{ tr('Your server name.') }}</div>
+                </div>
 
-            <div class="mdui-textfield mdui-textfield-floating-label">
-                <label class="mdui-textfield-label">{{ tr('Server IP') }}</label>
-                <input class="mdui-textfield-input" type="text" name="ip" value="{{ $server->ip }}" />
-            </div>
+                <div class="mdui-textfield mdui-textfield-floating-label">
+                    <label class="mdui-textfield-label">{{ tr('Server IP') }}</label>
+                    <input class="mdui-textfield-input" type="text" name="ip" value="{{ $server->ip }}" />
+                </div>
 
-            <div class="mdui-textfield mdui-textfield-floating-label">
-                <label class="mdui-textfield-label">{{ tr('Server Port') }}</label>
-                <input class="mdui-textfield-input" type="text" name="port" value="{{ $server->port }}" />
-            </div>
+                <div class="mdui-textfield mdui-textfield-floating-label">
+                    <label class="mdui-textfield-label">{{ tr('Server Port') }}</label>
+                    <input class="mdui-textfield-input" type="text" name="port" value="{{ $server->port }}" />
+                </div>
 
-            <div class="mdui-textfield mdui-textfield-floating-label">
-                <label class="mdui-textfield-label">Motd</label>
-                <input class="mdui-textfield-input" type="text" name="motd" value="{{ $server->motd }}" />
-            </div>
+                <div class="mdui-textfield mdui-textfield-floating-label">
+                    <label class="mdui-textfield-label">Motd</label>
+                    <input class="mdui-textfield-input" type="text" name="motd" value="{{ $server->motd }}" />
+                </div>
 
-            <div class="mdui-textfield mdui-textfield-floating-label">
-                <label class="mdui-textfield-label">Token</label>
-                <input class="mdui-textfield-input" type="text" value="{{ $server->token }}" readonly />
-            </div>
+                <div class="mdui-textfield mdui-textfield-floating-label">
+                    <label class="mdui-textfield-label">Token</label>
+                    <input class="mdui-textfield-input" type="text" value="{{ $server->token }}" readonly />
+                </div>
 
+                <button class="mdui-m-t-4 mdui-btn mdui-ripple mdui-btn-outlined"
+                    type="submit">{{ tr('Update Server') }}</button>
+            </form>
+
+        </div>
+
+        <div id="delete">
             <button class="mdui-m-t-4 mdui-btn mdui-ripple mdui-btn-outlined"
-                type="submit">{{ tr('Update Server') }}</button>
-        </form>
-
-    </div>
-
-    <div id="delete">
-        <button class="mdui-m-t-4 mdui-btn mdui-ripple mdui-btn-outlined"
-            onclick="m.delete()">{{ tr('Delete Server') }}</button>
-    </div>
+                onclick="m.delete()">{{ tr('Delete Server') }}</button>
+        </div>
+    @endif
 
     <script>
         m = {
@@ -101,6 +140,9 @@
                     util.delete(route('minecraftBeFlow.servers.destroy', {{ $server->id }}))
                 })
             },
+            // transfer: () => {
+            //     util.put(route('minecraftBeFlow.servers.update', {{ $server->id }}), $(ele).serializeArray())
+            // },
             e: (type, data) => {
                 switch (type) {
                     case 'minecraftBeFlow.server.updated':
